+++ src/database_api.jl
@@ -170,7 +170,7 @@ function pushStoppingConditionToDB(db_filepath::String, stopping_condition::Stop
     return stopping_condition_row_id
 end
 
-function pushSimulationToDB(db_filepath::String, sim_group_id::Union{Integer, Nothing}, prev_simulation_uuid::Union{String, Nothing}, db_id_tuple::NamedTuple{(:game_id, :graph_id, :sim_params_id, :starting_condition_id, :stopping_condition_id), NTuple{5, Int64}}, agent_graph::AgentGraph, periods_elapsed::Integer, distributed_uuid::Union{String, Nothing} = nothing)
+function pushSimulationToDB(db_filepath::String, sim_group_id::Union{Integer, Nothing}, prev_simulation_uuid::Union{String, Nothing}, db_id_tuple::NamedTuple{(:game_id, :graph_id, :sim_params_id, :starting_condition_id, :stopping_condition_id), NTuple{5, Int}}, agent_graph::AgentGraph, periods_elapsed::Integer, distributed_uuid::Union{String, Nothing} = nothing)
     #prepare simulation to be inserted
     adj_matrix_json_str = JSON3.write(Matrix(adjacency_matrix(agent_graph.graph)))
     rng_state = copy(Random.default_rng())
@@ -206,7 +206,7 @@ end
 
 
 function constructIDTuple(model::SimModel, db_filepath::String; use_seed::Bool = false)
-    db_id_tuple::NamedTuple{(:game_id, :graph_id, :sim_params_id, :starting_condition_id, :stopping_condition_id), NTuple{5, Int64}} = (
+    db_id_tuple::NamedTuple{(:game_id, :graph_id, :sim_params_id, :starting_condition_id, :stopping_condition_id), NTuple{5, Int}} = (
                     game_id = pushGameToDB(db_filepath, model.game),
                     graph_id = pushGraphToDB(db_filepath, model.graph_params),
                     sim_params_id = pushSimParamsToDB(db_filepath, model.sim_params, use_seed),
